# inputs/openapi.yaml
openapi: 3.1.0
info:
  title: Bookstore API
  description: REST API pro správu knihkupectví – knihy, autoři, kategorie a recenze.
  version: 1.0.0
paths:
...
  /authors:
    post:
      tags:
      - Authors
      summary: Create Author
      operationId: create_author_authors_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Authors
      summary: List Authors
      operationId: list_authors_authors_get
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorResponse'
                title: Response List Authors Authors Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /authors/{author_id}:
    get:
      tags:
      - Authors
      summary: Get Author
      operationId: get_author_authors__author_id__get
      parameters:
      - name: author_id
        in: path
        required: true
        schema:
          type: integer
          title: Author Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Authors
      summary: Update Author
      operationId: update_author_authors__author_id__put
      parameters:
      - name: author_id
        in: path
        required: true
        schema:
          type: integer
          title: Author Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Authors
      summary: Delete Author
      operationId: delete_author_authors__author_id__delete
      parameters:
      - name: author_id
        in: path
        required: true
        schema:
          type: integer
          title: Author Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /categories:
    get:
      tags:
      - Categories
      summary: List Categories
      operationId: list_categories_categories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoryResponse'
                type: array
                title: Response List Categories Categories Get
    post:
      tags:
      - Categories
      summary: Create Category
      operationId: create_category_categories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /categories/{category_id}:
    get:
      tags:
      - Categories
      summary: Get Category
      operationId: get_category_categories__category_id__get
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Categories
      summary: Update Category
      operationId: update_category_categories__category_id__put
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Categories
      summary: Delete Category
      operationId: delete_category_categories__category_id__delete
      parameters:
      - name: category_id
        in: path
        required: true
        schema:
          type: integer
          title: Category Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books:
    post:
      tags:
      - Books
      summary: Create Book
      operationId: create_book_books_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Books
      summary: List Books
      operationId: list_books_books_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: author_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Author Id
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Category Id
      - name: min_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Min Price
      - name: max_price
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            minimum: 0.0
          - type: 'null'
          title: Max Price
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBooks'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}:
    get:
      tags:
      - Books
      summary: Get Book
      operationId: get_book_books__book_id__get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Books
      summary: Update Book
      operationId: update_book_books__book_id__put
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Books
      summary: Delete Book
      operationId: delete_book_books__book_id__delete
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/reviews:
    post:
      tags:
      - Reviews
      summary: Create Review
      operationId: create_review_books__book_id__reviews_post
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Reviews
      summary: List Reviews
      operationId: list_reviews_books__book_id__reviews_get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
                title: Response List Reviews Books  Book Id  Reviews Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/rating:
    get:
      tags:
      - Reviews
      summary: Get Book Rating
      operationId: get_book_rating_books__book_id__rating_get
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/discount:
    post:
      tags:
      - Books
      summary: Apply Discount
      operationId: apply_discount_books__book_id__discount_post
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /books/{book_id}/stock:
    patch:
      tags:
      - Books
      summary: Update Stock
      operationId: update_stock_books__book_id__stock_patch
      parameters:
      - name: book_id
        in: path
        required: true
        schema:
          type: integer
          title: Book Id
      - name: quantity
        in: query
        required: true
        schema:
          type: integer
          title: Quantity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuthorCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        born_year:
          anyOf:
          - type: integer
            maximum: 2026.0
            minimum: 0.0
          - type: 'null'
          title: Born Year
      type: object
      required:
      - name
      title: AuthorCreate
    AuthorResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        born_year:
          anyOf:
          - type: integer
          - type: 'null'
          title: Born Year
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - bio
      - born_year
      - created_at
      title: AuthorResponse
    AuthorUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Name
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        born_year:
          anyOf:
          - type: integer
            maximum: 2026.0
            minimum: 0.0
          - type: 'null'
          title: Born Year
      type: object
      title: AuthorUpdate
...
    CategoryCreate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      title: CategoryCreate
    CategoryResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - id
      - name
      - description
      title: CategoryResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...